Reworking the page and layout structure, making SSR more useful
